Move failure_banner.dart and link_text_span.dart to others"
