+++ packages/clima_ui/lib/widgets/others/failure_banner.dart
